Optimize coverage, minor cleanups
