+++ .dockerignore
@@ -13,6 +13,7 @@ justfile
 **/*.rs.bk
 .idea/
 test_log*
+*.profraw
 
 pg_data/
 config.yml
