+++ Cargo.toml
@@ -1,4 +1,5 @@
 [workspace]
+default-members = ["martin-tile-utils", "."]
 members = ["martin-tile-utils"]
 
 [package]
