+++ src/pg/configurator.rs
@@ -6,10 +6,10 @@ use crate::pg::pg_source::{PgSource, PgSqlInfo};
 use crate::pg::pool::Pool;
 use crate::pg::table_source::{calc_srid, get_table_sources, merge_table_info, table_to_query};
 use crate::pg::utils::PgError::InvalidTableExtent;
-use crate::pg::utils::Result;
+use crate::pg::utils::{Result, Schemas};
 use crate::source::IdResolver;
 use crate::srv::server::Sources;
-use crate::utils::{find_info, normalize_key, InfoMap, Schemas};
+use crate::utils::{find_info, normalize_key, InfoMap};
 use futures::future::join_all;
 use itertools::Itertools;
 use log::{debug, error, info, warn};
