+++ tests/function_source_test.rs
@@ -2,8 +2,8 @@ use ctor::ctor;
 use itertools::Itertools;
 use log::info;
 use martin::pg::function_source::get_function_sources;
+use martin::pg::utils::Schemas;
 use martin::source::Xyz;
-use martin::utils::Schemas;
 
 #[path = "utils.rs"]
 mod utils;
